pyproject.toml
[tool.poetry]
name = "djproject"
version = "0.1.0"
description = ""
authors = ["wangdong <dong.wang3@bluefocus.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.5"
mysqlclient = "^2.0.3"
pydantic = "^1.7.3"
fastapi = "^0.63.0"
djangorestframework = "^3.12.2"
redis = "^3.5.3"
thrift = "^0.13.0"
itchat = "^1.3.10"
pygorithm = "^1.0.4"
mycli = "^1.23.2"
numpy = "^1.20.0"
scipy = "^1.6.0"
matplotlib = "^3.3.4"
pandas = "^1.2.1"
djangorestframework-simplejwt = "^4.6.0"
django-secret-key = "^1.0.1"
DingtalkChatbot = "^1.5.2"
sentry-sdk = "^1.0.0"
django-redis = "^4.12.1"
flower = "^0.9.7"
django-smart-selects = "^1.5.9"
setuptools-scm = "^6.0.1"
django-haystack = "^3.0"
pytest = "^6.2.2"
pytest-django = "^4.1.0"
uvicorn = "^0.13.4"
gunicorn = "^20.0.4"
uvloop = "^0.15.2"
httptools = "^0.1.1"
fire = "^0.4.0"
pybloom_live = "^3.0.0"
redisbloom = "^0.4.0"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
django-mysql = "^3.11.1"
Scrapy = "^2.4.1"
scrapy-djangoitem = "^1.1.1"



[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
